+++ gabarit/template_num/num_project/package_name/models_training/classifiers/models_sklearn/model_gbt_classifier.py
@@ -33,9 +33,9 @@ from sklearn.multioutput import MultiOutputClassifier
 from sklearn.ensemble import GradientBoostingClassifier
 from sklearn.multiclass import OneVsRestClassifier, OneVsOneClassifier
 
-from {{package_name}} import utils
-from {{package_name}}.models_training.model_pipeline import ModelPipeline
-from {{package_name}}.models_training.classifiers.model_classifier import ModelClassifierMixin  # type: ignore
+from .... import utils
+from ...model_pipeline import ModelPipeline
+from ..model_classifier import ModelClassifierMixin  # type: ignore
 
 
 class ModelGBTClassifier(ModelClassifierMixin, ModelPipeline):
