+++ gabarit/template_num/num_project/package_name/models_training/classifiers/models_sklearn/model_lgbm_classifier.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-## Logistic Regression model
+## Light GBM model
 # Copyright (C) <2018-2022>  <Agence Data Services, DSI Pôle Emploi>
 #
 # This program is free software: you can redistribute it and/or modify
@@ -17,7 +17,7 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 # Classes :
-# - ModelLogisticRegressionClassifier -> Logistic Regression mode for classification
+# - ModelLGBMClassifier -> Light GBM model for classification
 
 
 import os
@@ -29,28 +29,28 @@ import dill as pickle
 from typing import Union
 
 from sklearn.pipeline import Pipeline
-from sklearn.linear_model import LogisticRegression
+from lightgbm import LGBMClassifier
 from sklearn.multioutput import MultiOutputClassifier
 from sklearn.multiclass import OneVsRestClassifier, OneVsOneClassifier
 
-from {{package_name}} import utils
-from {{package_name}}.models_training.model_pipeline import ModelPipeline
-from {{package_name}}.models_training.classifiers.model_classifier import ModelClassifierMixin  # type: ignore
+from .... import utils
+from ...model_pipeline import ModelPipeline
+from ..model_classifier import ModelClassifierMixin  # type: ignore
 
 
-class ModelLogisticRegressionClassifier(ModelClassifierMixin, ModelPipeline):
-    '''Logistic Regression mode for classification'''
+class ModelLGBMClassifier(ModelClassifierMixin, ModelPipeline):
+    '''Light GBM model for Classification'''
 
-    _default_name = 'model_lr_classifier'
+    _default_name = 'model_lgbm_classifier'
 
-    def __init__(self, lr_params: Union[dict, None] = None, multiclass_strategy: Union[str, None] = None, **kwargs) -> None:
+    def __init__(self, lgbm_params: Union[dict, None] = None, multiclass_strategy: Union[str, None] = None, **kwargs) -> None:
         '''Initialization of the class (see ModelPipeline, ModelClass & ModelClassifierMixin for more arguments)
 
         Kwargs:
-            lr_params (dict) : Parameters for the Logistic Regression
-            multiclass_strategy (str): Multi-classes strategy, 'ovr' (OneVsRest), or 'ovo' (OneVsOne). If None, use the default of the algorithm.
-        Raises:
+            lgbm_params (dict) : Parameters for the Light GBM
             multiclass_strategy (str):  Multi-classes strategy, 'ovr' (OneVsRest), or 'ovo' (OneVsOne). If None, use the default of the algorithm.
+        Raises:
+            If multiclass_strategy is not 'ovo', 'ovr' or None
         '''
         if multiclass_strategy is not None and multiclass_strategy not in ['ovo', 'ovr']:
             raise ValueError(f"The value of 'multiclass_strategy' must be 'ovo' or 'ovr' (not {multiclass_strategy})")
@@ -61,24 +61,24 @@ class ModelLogisticRegressionClassifier(ModelClassifierMixin, ModelPipeline):
         self.logger = logging.getLogger(__name__)
 
         # Manage model
-        if lr_params is None:
-            lr_params = {}
-        self.lr = LogisticRegression(**lr_params)
+        if lgbm_params is None:
+            lgbm_params = {}
+        self.lgbm = LGBMClassifier(**lgbm_params)
         self.multiclass_strategy = multiclass_strategy
 
         # Can't do multi-labels / multi-classes
         if not self.multi_label:
             # If not multi-classes : no impact
             if multiclass_strategy == 'ovr':
-                self.pipeline = Pipeline([('lr', OneVsRestClassifier(self.lr))])
+                self.pipeline = Pipeline([('lgbm', OneVsRestClassifier(self.lgbm))])
             elif multiclass_strategy == 'ovo':
-                self.pipeline = Pipeline([('lr', OneVsOneClassifier(self.lr))])
+                self.pipeline = Pipeline([('lgbm', OneVsOneClassifier(self.lgbm))])
             else:
-                self.pipeline = Pipeline([('lr', self.lr)])
+                self.pipeline = Pipeline([('lgbm', self.lgbm)])
 
-        # LogisticRegression does not natively support multi-labels
+        # LGBMClassifier does not natively support multi-labels
         if self.multi_label:
-            self.pipeline = Pipeline([('lr', MultiOutputClassifier(self.lr))])
+            self.pipeline = Pipeline([('lgbm', MultiOutputClassifier(self.lgbm))])
 
     @utils.trained_needed
     def predict_proba(self, x_test: pd.DataFrame, **kwargs) -> np.ndarray:
@@ -181,9 +181,9 @@ class ModelLogisticRegressionClassifier(ModelClassifierMixin, ModelPipeline):
 
         # Manage multi-labels or multi-classes
         if not self.multi_label and self.multiclass_strategy is None:
-            self.lr = self.pipeline['lr']
+            self.lgbm = self.pipeline['lgbm']
         else:
-            self.lr = self.pipeline['lr'].estimator
+            self.lgbm = self.pipeline['lgbm'].estimator
 
         # Reload pipeline preprocessing
         with open(preprocess_pipeline_path, 'rb') as f:
