+++ gabarit/template_num/num_project/package_name/models_training/classifiers/models_sklearn/model_ridge_classifier.py
@@ -33,9 +33,9 @@ from sklearn.linear_model import RidgeClassifier
 from sklearn.multioutput import MultiOutputClassifier
 from sklearn.multiclass import OneVsRestClassifier, OneVsOneClassifier
 
-from {{package_name}} import utils
-from {{package_name}}.models_training.model_pipeline import ModelPipeline
-from {{package_name}}.models_training.classifiers.model_classifier import ModelClassifierMixin  # type: ignore
+from .... import utils
+from ...model_pipeline import ModelPipeline
+from ..model_classifier import ModelClassifierMixin  # type: ignore
 
 
 class ModelRidgeClassifier(ModelClassifierMixin, ModelPipeline):
