+++ gabarit/template_num/num_project/package_name/models_training/regressors/models_sklearn/model_kernel_ridge_regressor.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-## Stochastic Gradient Descent model
+## Kernel Ridge model
 # Copyright (C) <2018-2022>  <Agence Data Services, DSI Pôle Emploi>
 #
 # This program is free software: you can redistribute it and/or modify
@@ -17,7 +17,7 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 # Classes :
-# - ModelSGDRegressor -> Stochastic Gradient Descent model for regression
+# - ModelKernelRidgeRegressor -> Kernel Ridge model for regression
 
 
 import os
@@ -27,22 +27,22 @@ import dill as pickle
 from typing import Union
 
 from sklearn.pipeline import Pipeline
-from sklearn.linear_model import SGDRegressor
+from sklearn.kernel_ridge import KernelRidge
 
-from {{package_name}}.models_training.model_pipeline import ModelPipeline
-from {{package_name}}.models_training.regressors.model_regressor import ModelRegressorMixin  # type: ignore
+from ...model_pipeline import ModelPipeline
+from ..model_regressor import ModelRegressorMixin  # type: ignore
 
 
-class ModelSGDRegressor(ModelRegressorMixin, ModelPipeline):
-    '''Stochastic Gradient Descent model for regression'''
+class ModelKernelRidgeRegressor(ModelRegressorMixin, ModelPipeline):
+    '''Kernel Ridge model for regression'''
 
-    _default_name = 'model_sgd_regressor'
+    _default_name = 'model_kernel_ridge_regressor'
 
-    def __init__(self, sgd_params: Union[dict, None] = None, **kwargs) -> None:
+    def __init__(self, kernel_ridge_params: Union[dict, None] = None, **kwargs) -> None:
         '''Initialization of the class (see ModelPipeline, ModelClass & ModelRegressorMixin for more arguments)
 
         Kwargs:
-            sgd_params (dict) : Parameters for the Stochastic Gradient Descent
+            kernel_ridge_params (dict) : Parameters for the Kernel Ridge
         '''
         # Init.
         super().__init__(**kwargs)
@@ -51,11 +51,11 @@ class ModelSGDRegressor(ModelRegressorMixin, ModelPipeline):
         self.logger = logging.getLogger(__name__)
 
         # Manage model
-        if sgd_params is None:
-            sgd_params = {}
-        self.sgd = SGDRegressor(**sgd_params)
+        if kernel_ridge_params is None:
+            kernel_ridge_params = {}
+        self.kernel_ridge = KernelRidge(**kernel_ridge_params)
         # We define a pipeline in order to be compatible with other models
-        self.pipeline = Pipeline([('sgd', self.sgd)])
+        self.pipeline = Pipeline([('kernel_ridge', self.kernel_ridge)])
 
     def reload_from_standalone(self, **kwargs) -> None:
         '''Reloads a model from its configuration and "standalones" files
@@ -111,7 +111,7 @@ class ModelSGDRegressor(ModelRegressorMixin, ModelPipeline):
             self.pipeline = pickle.load(f)
 
         # Reload pipeline elements
-        self.sgd = self.pipeline['sgd']
+        self.kernel_ridge = self.pipeline['kernel_ridge']
 
         # Reload pipeline preprocessing
         with open(preprocess_pipeline_path, 'rb') as f:
