+++ gabarit/template_num/num_project/package_name/models_training/regressors/models_sklearn/model_pls_regressor.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-## K-nearest Neighbors model
+## Partial Least Squares model
 # Copyright (C) <2018-2022>  <Agence Data Services, DSI Pôle Emploi>
 #
 # This program is free software: you can redistribute it and/or modify
@@ -17,7 +17,7 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 # Classes :
-# - ModelKNNClassifier -> K-nearest Neighbors model for classification
+# - ModelPLSRegressor -> Partial Least Squares model for regression
 
 
 import os
@@ -29,30 +29,24 @@ import dill as pickle
 from typing import Union
 
 from sklearn.pipeline import Pipeline
-from sklearn.neighbors import KNeighborsClassifier
-from sklearn.multiclass import OneVsRestClassifier, OneVsOneClassifier
+from sklearn.cross_decomposition import PLSRegression
 
-from {{package_name}} import utils
-from {{package_name}}.models_training.model_pipeline import ModelPipeline
-from {{package_name}}.models_training.classifiers.model_classifier import ModelClassifierMixin  # type: ignore
+from .... import utils
+from ...model_pipeline import ModelPipeline
+from ..model_regressor import ModelRegressorMixin  # type: ignore
 
 
-class ModelKNNClassifier(ModelClassifierMixin, ModelPipeline):
-    '''K-nearest Neighbors model for classification'''
+class ModelPLSRegressor(ModelRegressorMixin, ModelPipeline):
+    '''Partial Least Squares model for regression'''
 
-    _default_name = 'model_knn_classifier'
+    _default_name = 'model_pls_regressor'
 
-    def __init__(self, knn_params: Union[dict, None] = None, multiclass_strategy: Union[str, None] = None, **kwargs) -> None:
-        '''Initialization of the class (see ModelPipeline, ModelClass & ModelClassifierMixin for more arguments)
+    def __init__(self, pls_params: Union[dict, None] = None, **kwargs) -> None:
+        '''Initialization of the class (see ModelPipeline, ModelClass & ModelRegressorMixin for more arguments)
 
         Kwargs:
-            knn_params (dict) : Parameters for the K-nearest Neighbors
-            multiclass_strategy (str):  Multi-classes strategy, 'ovr' (OneVsRest), or 'ovo' (OneVsOne). If None, use the default of the algorithm.
-        Raises:
-            ValueError: If multiclass_strategy is not 'ovo', 'ovr' or None
+            pls_params (dict) : Parameters for the Partial Least Squares
         '''
-        if multiclass_strategy is not None and multiclass_strategy not in ['ovo', 'ovr']:
-            raise ValueError(f"The value of 'multiclass_strategy' must be 'ovo' or 'ovr' (not {multiclass_strategy})")
         # Init.
         super().__init__(**kwargs)
 
@@ -60,63 +54,25 @@ class ModelKNNClassifier(ModelClassifierMixin, ModelPipeline):
         self.logger = logging.getLogger(__name__)
 
         # Manage model
-        if knn_params is None:
-            knn_params = {}
-        self.knn = KNeighborsClassifier(**knn_params)
-        self.multiclass_strategy = multiclass_strategy
-
-        # Can't do multi-labels / multi-classes
-        if not self.multi_label:
-            # If not multi-classes : no impact
-            if multiclass_strategy == 'ovr':
-                self.pipeline = Pipeline([('knn', OneVsRestClassifier(self.knn))])
-            elif multiclass_strategy == 'ovo':
-                self.pipeline = Pipeline([('knn', OneVsOneClassifier(self.knn))])
-            else:
-                self.pipeline = Pipeline([('knn', self.knn)])
-
-        # LKNeighborsClassifier natively supports multi_labels
-        if self.multi_label:
-            self.pipeline = Pipeline([('knn', self.knn)])
+        if pls_params is None:
+            pls_params = {}
+        self.pls = PLSRegression(**pls_params)
+        # We define a pipeline in order to be compatible with other models
+        self.pipeline = Pipeline([('pls', self.pls)])
 
     @utils.trained_needed
-    def predict_proba(self, x_test: pd.DataFrame, **kwargs) -> np.ndarray:
-        '''Predicts the probabilities on the test set
-            'ovo' can't predict probabilities : by default, return 1 for the predicted class, 0 otherwise.
+    def predict(self, x_test: pd.DataFrame, return_proba: bool = False, **kwargs) -> np.ndarray:
+        '''Predictions on test set. We override this function because the PLS does not return the same
+        format for the predictions.
 
         Args:
             x_test (pd.DataFrame): DataFrame with the test data to be predicted
-        Returns:
-            (np.ndarray): Array, shape = [n_samples, n_classes]
-        '''
-        # Uses super() of the ModelPipeline class if != 'ovo' or multi-labels
-        if self.multi_label or self.multiclass_strategy != 'ovo':
-            return super().predict_proba(x_test=x_test, **kwargs)
-        else:
-            # We check input format
-            x_test, _ = self._check_input_format(x_test)
-            # Get preds
-            preds = self.pipeline.predict(x_test)
-            # Format ['a', 'b', 'c', 'a', ..., 'b']
-            # Transform to "proba"
-            transform_dict = {col: [0. if _ != i else 1. for _ in range(len(self.list_classes))] for i, col in enumerate(self.list_classes)}
-            probas = np.array([transform_dict[x] for x in preds])
-        return probas
-
-    def save(self, json_data: Union[dict, None] = None) -> None:
-        '''Saves the model
-
         Kwargs:
-            json_data (dict): Additional configurations to be saved
+            return_proba (bool): If the function should return the probabilities instead of the classes (Keras compatibility)
+        Returns:
+            (np.ndarray): Array, shape = [n_samples,]
         '''
-        # Save model
-        if json_data is None:
-            json_data = {}
-
-        json_data['multiclass_strategy'] = self.multiclass_strategy
-
-        # Save
-        super().save(json_data=json_data)
+        return np.array([_[0] for _ in super().predict(x_test, return_proba, **kwargs)])
 
     def reload_from_standalone(self, **kwargs) -> None:
         '''Reloads a model from its configuration and "standalones" files
@@ -156,12 +112,6 @@ class ModelKNNClassifier(ModelClassifierMixin, ModelPipeline):
         # Load confs
         with open(configuration_path, 'r', encoding='{{default_encoding}}') as f:
             configs = json.load(f)
-        # Can't set int as keys in json, so need to cast it after reloading
-        # dict_classes keys are always ints
-        if 'dict_classes' in configs.keys():
-            configs['dict_classes'] = {int(k): v for k, v in configs['dict_classes'].items()}
-        elif 'list_classes' in configs.keys():
-            configs['dict_classes'] = {i: col for i, col in enumerate(configs['list_classes'])}
 
         # Set class vars
         # self.model_name = # Keep the created name
@@ -170,19 +120,15 @@ class ModelKNNClassifier(ModelClassifierMixin, ModelPipeline):
         self.trained = configs.get('trained', True)  # Consider trained by default
         # Try to read the following attributes from configs and, if absent, keep the current one
         for attribute in ['model_type', 'x_col', 'y_col', 'columns_in', 'mandatory_columns',
-                          'list_classes', 'dict_classes', 'multi_label', 'level_save',
-                          'multiclass_strategy']:
+                          'level_save']:
             setattr(self, attribute, configs.get(attribute, getattr(self, attribute)))
 
         # Reload pipeline model
         with open(sklearn_pipeline_path, 'rb') as f:
             self.pipeline = pickle.load(f)
 
-        # Manage multi-labels or multi-classes
-        if not self.multi_label and self.multiclass_strategy is not None:
-            self.knn = self.pipeline['knn'].estimator
-        else:
-            self.knn = self.pipeline['knn']
+        # Reload pipeline elements
+        self.pls = self.pipeline['pls']
 
         # Reload pipeline preprocessing
         with open(preprocess_pipeline_path, 'rb') as f:
