+++ gabarit/template_num/num_project/package_name/models_training/regressors/models_sklearn/model_svr_regressor.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-## Bayesian ridge model
+## Support Vector Regression model
 # Copyright (C) <2018-2022>  <Agence Data Services, DSI Pôle Emploi>
 #
 # This program is free software: you can redistribute it and/or modify
@@ -17,7 +17,7 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 # Classes :
-# - ModelBayesianRidgeRegressor -> Bayesian ridge model for regression
+# - ModelSVRRegressor -> Support Vector Regression model for regression
 
 
 import os
@@ -26,23 +26,23 @@ import logging
 import dill as pickle
 from typing import Union
 
+from sklearn.svm import SVR
 from sklearn.pipeline import Pipeline
-from sklearn.linear_model import BayesianRidge
 
-from {{package_name}}.models_training.model_pipeline import ModelPipeline
-from {{package_name}}.models_training.regressors.model_regressor import ModelRegressorMixin  # type: ignore
+from ...model_pipeline import ModelPipeline
+from ..model_regressor import ModelRegressorMixin  # type: ignore
 
 
-class ModelBayesianRidgeRegressor(ModelRegressorMixin, ModelPipeline):
-    '''Bayesian ridge model for regression'''
+class ModelSVRRegressor(ModelRegressorMixin, ModelPipeline):
+    '''Support Vector Regression model for regression'''
 
-    _default_name = 'model_bayesian_ridge_regressor'
+    _default_name = 'model_svr_regressor'
 
-    def __init__(self, bayesian_ridge_params: Union[dict, None] = None, **kwargs) -> None:
+    def __init__(self, svr_params: Union[dict, None] = None, **kwargs) -> None:
         '''Initialization of the class (see ModelPipeline, ModelClass & ModelRegressorMixin for more arguments)
 
         Kwargs:
-            bayesian_ridge_params (dict) : Parameters for Bayesian Ridge
+            svr_params (dict) : Parameters for the Support Vector Regression
         '''
         # Init.
         super().__init__(**kwargs)
@@ -51,11 +51,11 @@ class ModelBayesianRidgeRegressor(ModelRegressorMixin, ModelPipeline):
         self.logger = logging.getLogger(__name__)
 
         # Manage model
-        if bayesian_ridge_params is None:
-            bayesian_ridge_params = {}
-        self.bayesian_ridge = BayesianRidge(**bayesian_ridge_params)
+        if svr_params is None:
+            svr_params = {}
+        self.svr = SVR(**svr_params)
         # We define a pipeline in order to be compatible with other models
-        self.pipeline = Pipeline([('bayesian_ridge', self.bayesian_ridge)])
+        self.pipeline = Pipeline([('svr', self.svr)])
 
     def reload_from_standalone(self, **kwargs) -> None:
         '''Reloads a model from its configuration and "standalones" files
@@ -111,7 +111,7 @@ class ModelBayesianRidgeRegressor(ModelRegressorMixin, ModelPipeline):
             self.pipeline = pickle.load(f)
 
         # Reload pipeline elements
-        self.bayesian_ridge = self.pipeline['bayesian_ridge']
+        self.svr = self.pipeline['svr']
 
         # Reload pipeline preprocessing
         with open(preprocess_pipeline_path, 'rb') as f:
