+++ gabarit/template_num/num_project/package_name/monitoring/mlflow_logger.py
@@ -30,7 +30,7 @@ import matplotlib.pyplot as plt
 import pandas as pd
 from typing import Union
 
-from {{package_name}} import utils
+from .. import utils
 
 # Deactivation of GIT warning for mlflow
 os.environ["GIT_PYTHON_REFRESH"] = "quiet"
