+++ gabarit/template_num/num_project/package_name/monitoring/model_explainer.py
@@ -29,8 +29,8 @@ import pandas as pd
 import matplotlib.pyplot as plt
 from typing import Type, Union, Any
 
-from {{package_name}}.models_training import utils_models
-from {{package_name}}.models_training.model_class import ModelClass
+from ..models_training import utils_models
+from ..models_training.model_class import ModelClass
 
 
 class Explainer:
